requirements.txt
# GitGlow Requirements
# Core dependencies for GitGlow GitHub LED matrix display

# LED Control
rpi-ws281x>=4.3.0          # WS2812B LED strip control for Raspberry Pi

# Web Framework  
flask>=2.3.0               # Web interface framework
hypercorn>=0.14.0          # ASGI server for async Flask apps
quart>=0.19.0              # Async Flask alternative

# HTTP Client
requests>=2.31.0           # HTTP requests for GitHub API
aiohttp>=3.8.0             # Async HTTP client

# Configuration
PyYAML>=6.0                # YAML configuration files
python-dotenv>=1.0.0       # Environment variable loading

# Network Management
wifi>=0.3.8                # WiFi management utilities
netifaces>=0.11.0          # Network interface information

# System Integration
systemd-python>=235        # Systemd service integration (Linux only)

# Development Dependencies
pytest>=7.4.0             # Testing framework
pytest-asyncio>=0.21.0    # Async testing support
black>=23.7.0              # Code formatting
flake8>=6.0.0              # Code linting
mypy>=1.5.0                # Type checking

# Hardware Simulation (for development without Pi)
fake-rpi>=0.7.0            # GPIO simulation for development

# Optional: Advanced LED Effects
numpy>=1.24.0              # Numerical operations for LED effects
Pillow>=10.0.0             # Image processing for graphics